do not promote Symbol keys over String keys.
it is odd that we would ask the user to create a Symbol for what will
be turned into a String. it allocates a hard-to-GC Symbol and a String,
when only a String is required.

accessing via Symbol still works, but we should promote the efficient option
over the inefficient one.