fix syntax hightlighting of change_prompt.rb
let's try to avoid exotic heredoc syntax that GitHub and multiple
editors cannot support, the only confirmed editor that supports
highlighting this syntax so far is 'emacs'.

it is not friendly to non-emacs users to have to deal with broken
syntax highlighting, especially when there is a less exotic and
easier syntax available.